+++ test/dummy/app/models/referer.rb
@@ -1,0 +1,3 @@
+class Referer < ApplicationRecord
+  belongs_to :user
+end